refactor: List 조회 시 아이템 순위대로 정렬하여 반환하도록 리팩터링 (#183)
