+++ src/main/java/com/listywave/collection/application/dto/CollectionResponse.java
@@ -23,70 +23,70 @@ public record CollectionResponse(
                 .map(CollectionListsResponse::of)
                 .toList();
     }
-}
 
-record CollectionListsResponse(
-        Long id,
-        ListsResponse list
-) {
+    public record CollectionListsResponse(
+            Long id,
+            ListsResponse list
+    ) {
 
-    public static CollectionListsResponse of(Collect collect) {
-        return new CollectionListsResponse(collect.getId(), toResponse(collect.getList()));
+        public static CollectionListsResponse of(Collect collect) {
+            return new CollectionListsResponse(collect.getId(), toResponse(collect.getList()));
+        }
+
+        public static ListsResponse toResponse(ListEntity list) {
+            return ListsResponse.of(list);
+        }
     }
 
-    public static ListsResponse toResponse(ListEntity list) {
-        return ListsResponse.of(list);
+    @Builder
+    public record ListsResponse(
+            Long id,
+            String backgroundColor,
+            String title,
+            Long ownerId,
+            String ownerNickname,
+            String ownerProfileImageUrl,
+            LocalDateTime updatedDate,
+            List<ListItemsResponse> listItems
+    ) {
+
+        public static ListsResponse of(ListEntity list) {
+            return ListsResponse.builder()
+                    .id(list.getId())
+                    .backgroundColor(list.getBackgroundColor())
+                    .title(list.getTitle().getValue())
+                    .ownerId(list.getUser().getId())
+                    .ownerNickname(list.getUser().getNickname())
+                    .ownerProfileImageUrl(list.getUser().getProfileImageUrl())
+                    .updatedDate(list.getUpdatedDate())
+                    .listItems(toList(list.getTop3Items().getValues()))
+                    .build();
+        }
+
+        public static List<ListItemsResponse> toList(List<Item> items) {
+            return items.stream()
+                    .sorted(Comparator.comparing(Item::getRanking))
+                    .map(ListItemsResponse::of)
+                    .limit(3)
+                    .toList();
+        }
     }
-}
 
-@Builder
-record ListsResponse(
-        Long id,
-        String backgroundColor,
-        String title,
-        Long ownerId,
-        String ownerNickname,
-        String ownerProfileImageUrl,
-        LocalDateTime updatedDate,
-        List<ListItemsResponse> listItems
-) {
+    @Builder
+    public record ListItemsResponse(
+            Long id,
+            int rank,
+            String title,
+            String imageUrl
+    ) {
 
-    public static ListsResponse of(ListEntity list) {
-        return ListsResponse.builder()
-                .id(list.getId())
-                .backgroundColor(list.getBackgroundColor())
-                .title(list.getTitle().getValue())
-                .ownerId(list.getUser().getId())
-                .ownerNickname(list.getUser().getNickname())
-                .ownerProfileImageUrl(list.getUser().getProfileImageUrl())
-                .updatedDate(list.getUpdatedDate())
-                .listItems(toList(list.getItems().getValues()))
-                .build();
-    }
-
-    public static List<ListItemsResponse> toList(List<Item> items) {
-        return items.stream()
-                .sorted(Comparator.comparing(Item::getRanking))
-                .map(ListItemsResponse::of)
-                .limit(3)
-                .toList();
-    }
-}
-
-@Builder
-record ListItemsResponse(
-        Long id,
-        int rank,
-        String title,
-        String imageUrl
-) {
-
-    public static ListItemsResponse of(Item item) {
-        return ListItemsResponse.builder()
-                .id(item.getId())
-                .rank(item.getRanking())
-                .title(item.getTitle().getValue())
-                .imageUrl(item.getImageUrl().getValue())
-                .build();
+        public static ListItemsResponse of(Item item) {
+            return ListItemsResponse.builder()
+                    .id(item.getId())
+                    .rank(item.getRanking())
+                    .title(item.getTitle().getValue())
+                    .imageUrl(item.getImageUrl().getValue())
+                    .build();
+        }
     }
 }
